Add EngineOutputFilter, filter by check name
Engines are instantiated with a string a json that corresponds to a
modified version of the YAML object under the engines key for the given
engine. Within this object, we'll now support a "checks" key that can
hold check-specific configuration including an enabled boolean.

This commit introduces an EngineOutputFilter that accepts this engine
configuration and decides if the output of the engine should be written
to the formatter.

There is a lot of JSON generation and re-parsing going on that can seem
redundant; I'm leaving this in place for now because orchestrating the
various parts that would need to change to address that is difficult.

One possible approach would be to enrich the interfaces involved and not
pass strings around so much. Engine would take a Hash of configuration
and stringify it only for the purposes of writing the config.json file.
Formatter could lose it's IO-like write interface and receive more
structured data centralizing the output parsing and filtering further
upstream.