Support container logging within Engine
We want to log the actions of starting, finishing, or timing out
containers into the database via Kafka/Wally. We will initially expose
this information in Admin and eventually to users via some kind of
Builds dashboard.

Unfortunately, these actions occur very far away from the code that is
actually able to perform the logging:

    CC::Builder  |  CC::Analyzer
    =============|============================
    Runner      -|-> EnginesRunner -> Engine
      |                                 |
      V                                 V
    Kafka                           docker run

The approach implemented here passes a ContainerLogger object into
EnginesRunner and onto each Engine. This object responds to event
methods which the Engine invokes as it starts and stops containers.

Within this code base, the object is a NullContainerLogger and does
nothing for these events, but Builder will be able to pass a real
ContainerLogger that writes the information to the database through
Kafka.

Other approaches considered and ruled out or deferred:

- Reifying a Container object to hold the docker run and (optionally)
  event logging logic

  This class could have been reusable for the (eventually
  container-based) Clone logic in Builder and taken advantage of the
  homogeneity that provides. Unfortunately, we would still be invoking
  Container from within Engine as it's very useful to share the
  EngineRunner and Engine objects between Builder and Analyzer, so it
  wouldn't help us in our immediate goal.

  When we do have the cloning performed in a container, creating such a
  class to hold the Posix::Spawn and timeout logic would make sense.

- Customizable life-cycle methods on Engine

  This would bloat Engine with new concerns and not be significantly
  more convenient for the objects that wanted to do logging.

  The object in charge of the actions to take only knows about
  EnginesRunner, so these actions would need to be passed through to
  Engine somehow.

  This would mean treating functions as values (setting and calling
  lambdas for the container life cycle events), and that is not
  idiomatic Ruby.