Pass container image and name to all logged events
More than just the start event needs to know about container-specific
information (so far the container image and name). One solution to this
would be to instantiate a new log per container and pass in those values
at construction. This does not work for the following reasons:

- The object instantiating the log does not know these values
- The log also needs to know *build*-specific values (e.g. host and
  snapshot), so we really want to instantiate a single log for the build
  with these values passed in at construction (and the constructing
  object does know these)

The solution chosen is to pass in the container-specific values for
every life-cycle event, not just started.