Add severity reporting
Many static analysis tools include a severity level
for each security warning they report.

This PR expects that levels can be one of three:
info < normal < critical

and updates the plaintext formatter to provide this
information if the engine provides it. It is not required.

No change is needed in the json formatter to allow the
severity key and description to be included in output.