spec helper: make YAML safe mode default
Because specs are run in a random order, there's a random occurence of
SafeYAML complaining about safe mode not being on when running tests,
because a spec calling `YAML.load` happens to get called before any
specs that load the runner (where this mode is set in normal execution).
This is annoying.