Extract Analyzer::Engines from Analyzer::EnginesRunner.
This reduces the responsibility of Analyzer::EnginesRunner so it only
runs analyses on an array of pre-instantiated Engines. Choosing and
instantiating engines with the correct config is now the responsibility
of Analyzer::Engines.

This is motivated in part by desired changes in stream handling in
platform -- separating these responsibilities here will give us more
flexibility elsewhere. In addition, the hope is that this might lead to
phasing out EnginesRunner entirely.