engine builder: handle nil engines in config
With the cc/yaml changes, an empty engines config is a yaml error, so
the node is nil. Since I guess we're not doing explicit validation for
of config for commands like this first, it seems prudent right now to
just ensure it will get a usable value & continue to do what it did
before.