Enforce a maximum output byte count on containers
Currently containers can generate an unbounded amount of output on
STDOUT and STDERR while running. This is bad for a variety of reasons,
but the symptom most pressing is that Docker has a memory leak around
dealing with large STDOUT streams especially when there are "slow"
readers attached.

This implements a concept of maximum output bytes that when tripped will
stop the container. The maximum output bytes is defaulted to 500MB.