Replace recursive directory class with minimizer
This replaces the existing Directory class with a minimizer class that
takes the paths to check against and diffs it against all the files in
project to get directories where possible instead of a list of files.

This change is necessary to be able to use engine specific excludes
without having to build up the include_paths over and over for each
engine.