Update to emit .codeclimate.yml missing message as warning, and return 0 exitcode instead of 1
This behavior is less confusing than returning a 1 exitcode. Since the init
command can add value to a repo by generating default configs where applicable,
we should simply emit a warning for the lack of .codeclimate.yml here.

It seems to me that this behavior change in `codeclimate` gem should only impact
CLI users, since Builder doesn't try to run `codeclimate init` to generate a config
if there is a .codeclimate.yml config present.

Config Inferrer on .com only has an impact when the .codeclimate.yml is missing or only has languages key.
https://github.com/codeclimate/builder/blob/master/lib/cc/builder/analysis_config_inferrer.rb#L7

codeclimate init --upgrade is used to transform a classically configured .codeclimate.yml into a config
for engines analysis.

At the moment, the --upgrade argument isn't v user-faceing (it's not listed in the --help
commands).

Although the `init` step checks for an edge case where a user has engines AND the
--upgrade flag has been set - and then complains noisily - we should not worry about
guarding against that case for the CLI. It shouldn't occur on .com, but if it does - failing loudly makes sense.