Fix analysis path bugs
This re-introduces `requested_paths` to several files that gets passed
into `include_paths_builder`. This was removed in an earlier commit but
is necessary for specifying individual files or paths to analyze instead
of analyzing the entire directory.

This also changes the PathMinimizer to check individual directory
entries to determine if it can minimize the paths or not.