Container: fix race with IO processing
There's a race between the main thread getting to this point after the
process finishes, and the two threads that process output from stdout &
stderr. When main thread "wins" that race, the result can be an
incorrect status for the events data on the engine step down the line:
e.g. the engine emits invalid output, but by the time the background
thread has tried to parse it & emitted an error, the main thread has
already decided the container is finished, so the error is never
attached to the relevant step.