Container: faster race resolution
This is an optimization of the fix from 320eda4d.

The previous code in a tight loop using `sleep` seemed like it thrashed
CPU context switches. Hence why it could potentially take quite a lot
longer, as evidenced by the spec change.

This should have the same behavior: `Thread#value` is a blocking call
that waits until the thread has finished. But this avoids the `sleep` on
the main thread which resulted in thread thrashing.

I have also devised a spec that actually tests this behavior reasonably
well. It's a bit hacky in some ways.