close io pipes when threads exit
This seems like a possible cause of hanging builders after timeouts: we
aren't manually closing these IO pipes, and so they may be left open
when the process they're connected to is killed on some platforms? The
`pspawn` docs even say you should be sure to close these.

I haven't been able to reproduce the hanging builders locally, so this
is a bit of a shot in the dark.