Fix output reader behavior in the face of timeouts
Previous behavior:

- Engine would run and complete in minutes
- `Process.waitpid` would return a successful status
- The else branch of if `@timed_out` would be followed
- If the engine produced a significant amount of stdout, and we're slow
  to read it, the output buffers in the docker process (bloating RAM)
  and we block in the `Thread#join` calls
- The timeout thread eventually reaches 15 minutes and invokes
  `#reap_running_container` directly (not `#stop`)
- The docker kill is a no-op because the container's already finished
- Because it didn't go via `#stop`, `@on_output` is left untouched and
  continues to churn

In such cases, the builder runs for as long as it takes to get through
the output. Things eventually do finish, but it takes forever and the
memory bloat of the buffered stdout can cause operational issues.

The fix is to ensure the stdout/stderr processing is governed by the
timeout thread. For this to function correctly, we had to move a few
things around:

- When we intentionally stop things, always go through `#stop`
- Within `#stop`, kill the output readers then and there
- `Thread#join` the output readers *before* checking `@timed_out`

In addition, we now pass a `message` argument through to the log entry
written when we kill a container.