Don't enable duplication by default
By surfacing & no longer silently swallowing errors that could lead to non-deterministic results, we are now in a bit of a tough spot because this is surfacing errors that look like straight-up bugs in the babel parser, but which programmatically aren't very distinguishable from previously mentioned non-deterministic results bugs.

(The underlying errors are parsing timeouts: previously these were low at 20 seconds and being swallowed, which was bad because memory pressure/gc behavior could swing execution to one side or the other of that line. Now they are high at 5 minutes & fatal, but due to what look like babel issues reliably happen on some files. It might be reasonable to swallow these extreme timeouts, but it would only take 2 or 3 of them to timeout the whole engine, so it hardly seems worthwhile.)