Drastically simplify workspace building
This commit removes IncludePathsBuilder and all classes involved in it.
In its place, we have CC::Workspace (new terminology) with drastically
simpler semantics. This simpler approach should be faster and more
reliable, at the expense of less features.

We drop support for glob-style patterns. Instead we support a simple
list of files or directories to ignore. We stop sending an exclude_paths
configuration key in /engines.json, only include_paths now. We stop
incorporating .gitignore in workspace-building.

Reasoning:

- Paths, not patterns.

  We see little actual value in patterns support. We leave it up to the
  engine to know that it should (for example) only target .py files,
  therefore we don't have to exclude **/*.pyc from the workspace.

  We've found the vast majority of our users have just been trying to
  exclude directories. They're actually quite frustrated having to
  shoe-horn it into patterns, only to be met with slowness or frequent
  bugs as we needlessly expand and filter these patterns in complicated
  ways.

  Old:

      exclude_paths:
      - .node_modules/**/*
      - app/assets/vendor/**

  New:

      exclude_paths:
      - .node_modules
      - app/assets/vendor

  We minimize disruption to these users by transparently handling a
  pattern such as foo/bar/** or foo/bar/**/** as the directory foo/bar.
  Other patterns, such as foo/*.pyc will generate a warning and be
  ignored.

- Ignore .gitignore

  The motivation for this feature is unclear, so I'm testing the waters
  about dropping it as it simplifies things.

- Stop all the permissions checking

  If the current directory contents, plus the user's defined excludes
  causes a permissions error in the workspace, it should be treated as
  an actual error: the user must modify the permissions or add the
  unreadable file to excludes.

Additional feature:

- Per-engine excludes

  The algorithm supports repeated filtering. That is, we can build a
  workspace filtered from top-level exclude_paths once, then filter that
  further for each engine with per-engine-excludes.

Caveats:

The "explicit includes" feature (i.e. codeclimate analyze [FILE, ...])
has an interesting interaction with this approach. Specifying files in
this way effectively means we begin building a workspace from these
files rather than the current directory's contents. This means excludes
will still apply on top of the passed files. This may not be what we
want, an alternative would be to pass along those paths as-is as
include_paths and do no filtering -- I leave the decision open, with no
strong opinion of my own.

Because the passed files are compared against the exclude list, they
must be simple and relative (i.e. they cannot be "." or "./"-prefixed),
as they would come out in a * glob. This is brittle, but currently
necessary. See PathsValidator for more details.

Additionally, the way the exclude list applies on top of the given files
can be surprising. A user could exclude spec, for example, and pass
spec/foo.rb. In such a case, spec/foo.rb will be processed which may be
surprising. This is because the current exclusion check assumes we'd
never even get to a file within spec if that directory were entirely
excluded, so there's no logic to catch such a case.

I'm inclined to either simplify the feature to a simple pass-through or
keep these caveats and just document the behavior.