Add Workspace::PathTree for path building
This is a new take on how to build up `include_paths` performantly while
still maintaining support for globs: a `PathTree` represents all files
of interest as a sparse tree, where each node is either a directory
(represented as a `PathTree` instance), or a plain file (a `String`).

The `PathTree` starts with the assumption that all files should be
included, and hence does not populate child nodes until it needs to in
order to calculate exclusions. As a result, directories are not walked
until they have to be, and a `PathTree`'s `#all_values` is a naturally
minimized set of paths of interest.