Add Workspace::PathTree for path building
This is a new take on how to build up `include_paths` performantly while
still maintaining support for globs: a `PathTree` represents all files
of interest as a lazily populated tree, where each node is either a
directory or a plain file.

The `PathTree` starts with the assumption that all files should be
included, and hence does not populate child nodes until it needs to in
order to calculate exclusions or explicit inclusion. As a result,
directories are not walked until they have to be, and a `PathTree`'s
`#all_values` is a naturally minimized set of paths of interest.

`Workspace` uses `PathTree` for including & excluding paths. This
restores glob support from the previous commit.

`PathsValidator` has been removed: it was fairly elementary to make
`PathTree` handle "./" prepended paths, which was the primary thing this
validator was there to prevent.

`.git` is always excluded explicitly by `EnginesConfigBuilder`. However,
like Pat explained in the previous commit, we are no longer making any
use of `.gitignore` patterns. Users should add paths to
`exclude_patterns` in their `.codeclimate.yml` to reflect this change if
necessary.

A new `rake benchmarks` task is available, which currently runs one spec
benchmarking the performance of the new `PathTree` implementation: the
previous poor performance was the result of a series of well-intentioned
changes which led to a pretty terrible performance situation, so it's
worth having an automated sanity check that things don't get terrible
again.