Enable the kibit engine
Although the kibit Code Climate engine is referenced in the online
documentation, it is not enabled in the list of engines. (*´･Д･)

As a Docker image for the engine hasn't been published to Docker Hub yet, which
may be why it isn't listed in the [list of community supported
engines](https://docs.codeclimate.com/docs/list-of-engines).

This change prempts the publishing of this Docker image and general
availability of the `codeclimate-kibit` engine. ( ᐛ )و