Init: add output for steps
On large repos, `init` can take long enough to scan for engines that it
can feel hung/unresponsive since it outputs nothing until after it's
calculated the full `.codeclimate.yml`.

This adds a brief message to let the user know that work is happening.

I also added a status message about generating the default config for
engines: we output messages if we *don't* generate a config for an
engine because the file already exists, but we never tell the user we
*are* going to generate configuration for any engines: without that
context, telling them we decided not to copy something  we never mentioned
we might copy in the first place is odd.

FWIW I tried to user `Analyzer::Spinner` here so the status line could
get some nice animation. Something within `create_codeclimate_yaml` must
be heavily blocking on IO or something, though, because the thread to
animate the spinner was getting woken very sporadically, resulting in a
very slow & unreliable spin animation. For now I think it's preferable
to have static text.