Init: respect excludes when enabling engines
`ConfigGenerator` now has a slightly improved algorithm for finding
files matching engine's `enable_regexps`. This fixes a bug & noticeably
improves performance in some cases as well.

The bug fix is that because we used to check the *entire* tree for files
matching an engine's patterns, you could enable an engine whose only
matching files were in auto-exclude paths. E.g. a project under
`vendor/` might have some JavaScript files used for local scripting or
something while your own code had no JS at all.

The performance benefit comes for similar reasons to why the recent
refactor of workspace calculation with exclude paths was an improvement:
we touch less data && less of the filesystem.

The new algorithm is similar to the old one: we still shell out to
`find` for available files, but `find` now only searches files that are
*NOT* within the excluded directories. I tried a variety of
implementations here, including things more similar to (or using) the
new `PathTree` implementation, but this was still the fastest
implementation across a variety of cases.

I also made a behavior change here: this won't auto-enable any engine
based on files in "." directories: this is an opinion I'm happy to back
away from, but in general "." directories & files are usually stuff not
related to your actual code that you would want analyzed, so this seems
reasonable. It does introduce a slight gap in behavior in that we don't
exclude those directories from analysis, just from enabling engines.

This also removes unused functionality from `Analysis::Filesystem`.