Init: don't create engine configs when inappropriate
Because of the behavior where we create engine configs even if a
.codeclimate.yml exists, there was a case where
`#create_default_engine_configs` would do the wrong thing: that method
still called `#available_engines` on `ConfigGenerator` to decide which
engines to create configs for. However, in the case of a
`.codeclimate.yml` created or edited by the user, there could be engines
that available_engines considers available, but that are not enabled in
the yaml. In that case, this method would create config files for
engines the user isn't using.

This changes the behavior to read the engines from the
`.codeclimate.yml`, since it must exist by the time this method is
called. As a bonus, in cases where the `.codeclimate.yml` already
existed & we didn't generate it, reading this from the yaml can be much
faster than using the `ConfigGenerator` logic.