ConfigGenerator: Catch & report `find` failure
Our usage of `find` to get files to enable engines against is too
naive: there are some unusual local configurations that lead to an
invalid find invocation. The old code was effectively ignoring this,
leading to more confusing failures further down the line.

This replaces the naive execution with usage of `POSIX::Spawn` that
should 1) be more tolerant of weird file names anyway and 2) if find
does fail, will properly catch & report so at least the user gets an
error message pointing them in a useful direction.