Refactor handling of mounted code and tmp paths
Adds a "MountedPath" abstraction that provides a single identity for the
case where we have an in-container path and need to know the on-host
path it represents for the purpose of re-mounting into other containers.

Removes the hap-hazard FILESYSTEM_DIR and inconsistently-named CODE_PATH
environment variables. FILESYSTEM_DIR goes away as MountedPath can
maintain that functionality itself. CODE_PATH is replaced with
CODECLIMATE_CODE to be consistent with the new CODECLIMATE_TMP and
existing CODECLIMATE_DEBUG.

We continue to read CODE_PATH if CODECLIMATE_CODE is unspecified. This
is to support any existing users with their own wrapper scripts in the
wild (interestingly, Builder fits in this category).

MountedPath internally makes use of a CODECLIMATE_DOCKER variable to set
the local and in-container paths as equal if not being run in a
container. This makes local development and testing more convenient.

Having CODECLIMATE_TMP means we don't have to force a root-owned /tmp/cc
on our users' machines. Users can set this variable to whatever they
want. We make use of this in the specs.

Builder should work as-is. It assumes /tmp/cc, but that was maintained
as the default when CODECLIMATE_TEMP is not set. We would now be free to
use that variable to make a build-specific temporary directory though.

One unknown: the test subcommand was messing with the CODE_PATH and
FILESYSTEM_DIR variables as part of running its analysis. This code is
untested (ha) and I don't understand it, so I'm not sure what needs to
change to support the MountedPath approach.