PathTree: fix excluding of directories directly
There was a bug here where if a directory was populated (probably due to
being partially excluded), and then later completely excluded (such as
for an engine-specific exclude), it would not actually be fully removed.

The fix here is that the check for "remove the node completely" now
checks that the child is not populated, or that it is the last thing
asking to be removed (the tail is empty).