Install codeclimate-wrapper from image
The Anywhere instructions run make install from a source archive of the
master branch. If we make a change to the wrapper script in master but
not (yet) supported by the released codeclimate/codeclimate image, this
can cause problems. Installing the wrapper out of the released image
ensures it's always the one meant to be used with that image.