Validate that issue-path is a file
Emitting issues for directories is not something we currently support.
Doing so was not causing visible problems until we added source-based
fingerprinting which attempts to read the path and errors on
directories:

https://bugsnag.com/code-climate/builder/errors/5736508dbe3f29e6a6c1cde8?event_id=573e507dba76e62611aa1445

Note: this still returns true for symlinks; discussion for how to handle
those cases is still open and this patch should not impact that behavior
either way.