Improve error messaging on ImageRequired
Previously, this would only be caused by our own screw up in the
manifest file. This meant a more spartan message and reliance on CX was
fine. Now, a user could have configured a bunk channel so we'd like to
guide them for how to fix it.

Analyzer::Engine is the best place to provide that context so we rescue
and re-raise with a better message. Doing it there is CLI-only, so it'll
need to be done again in Builder, but the behavior will be different
(we'll need an error code, etc) so that's expected and appropriate.