Fix README help for `analyze` command
The docs as written-here were confusing. For one thing, they were incorrectly
implying that `path` is a required argument: it's not, and providing a path like
`.` has suprising behavior for a lot of users.

This updates the short-help to make `path` more obviously optional, and adds a
paragraph to the fuller description below explaining the behavior of `analyze`
more fully both when `path` is provided & when it is not.