Cache the other_location check by path.
This speeds up anything that makes use of other_locations extensively.
If one location is valid, all of them (with the same path) will be.
This took codeclimate-duplication on big_files from 603s to 108s.

More could probably be done to speed up the driver, but this was
definitely at the top of my profile.