Exit 1 when due to an unexpected error
We exit 0 in `Runner` when handling exceptions. That's pretty surprising, and is
very problematic for any attempt to use the CLI within a script or an editor
plugin or anything of that kind: it can lead to accidentally think everything
went fine & found no issues when in fact analysis just didn't run. I'm shocked
we've made it this long without this being a problem, actually.