Support a #reload operation on Config objects
The primary use-case is the prepare.files step and eventual auto-enable plugins
feature. If we load and validate a config, pull in a .rubocop.yml through
prepare.files, we want Config#engines to auto-enable rubocop because that file
is now present. To do this we'll Config#reload after prepare.

Note: the reload does not re-read the yaml from disk. This is to avoid the
caller forgetting to re-validate a reloaded config, and risking confusing
downstream errors. That means we don't support prepare.files pulling down a
change to codeclimate.yml itself.